requirements.txt
streamlit
pandas
numpy
plotly
python-binance
scikit-learn
tensorflow-cpu
xgboost
feedparser
transformers
torch
requests